pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolio-ai"
version = "0.1.0"
description = "AI-powered investment portfolio analysis tool"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "ecoto"}
]
dependencies = [
    "pandas>=2.2",
    "numpy>=2.0",
    "yfinance>=0.2.40",
    "openpyxl>=3.1",
    "anthropic>=0.40",
    "click>=8.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
]

[project.scripts]
portfolio-ai = "portfolio_ai.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=portfolio_ai --cov-report=term-missing"

[tool.coverage.run]
source = ["src/portfolio_ai"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
